Copied code for Task 3
Visualized the load paths for task 3
For that reworked evaluation.py an called it data_visualization.py and
changed read_txt method when reading W

+++ 02_hyperelasticity_I/modules/models.py
@@ -188,7 +188,7 @@ main: construction of the NN model
 def main(loss_weights, **kwargs):
     """ This creates a Keras model """
     # define input shape
-    xs = tf.keras.Input(shape=(3,3), dtype='float32')
+    xs = tf.keras.Input(shape=(3,3))
     # define which (custom) layers the model uses
     l_nn = make_layer(**kwargs)
     ys = l_nn(xs)
